Add file and directory check

+++ Makefile
@@ -3,7 +3,7 @@
 all: updateVim
 
 updateVim:
-	mv $${HOME}/.vimrc $${HOME}/.vimrc.bak
+	[ -f $${HOME}/.vimrc ] && mv $${HOME}/.vimrc $${HOME}/.vimrc.bak
 	cp $${HOME}/.quaren/.vimrc $${HOME}/.vimrc
-	mv $${HOME}/.vim $${HOME}/.vim.bak
+	[ -d $${HOME}/.vim ] && mv $${HOME}/.vim $${HOME}/.vim.bak
 	cp -r $${HOME}/.quaren/.vim $${HOME}/.vim
